move handler struct and config impls to handler.rs, abstract trie layout
